Removed heads that are not implemented in tensorflow 2
