+++ rl_coach/architectures/tensorflow_components/losses/v_loss.py
@@ -16,10 +16,11 @@
 import tensorflow as tf
 from tensorflow import keras
 from tensorflow.keras.losses import Loss, Huber, MeanSquaredError
+from rl_coach.architectures.tensorflow_components.losses.head_loss import HeadLoss, LossInputSchema
 
 from rl_coach.architectures.mxnet_components.heads.head import LOSS_OUT_TYPE_LOSS
 
-class VLoss(keras.losses.Loss):
+class VLoss(HeadLoss):
 
     def __init__(self, network_name,
                  head_idx: int = 0,
@@ -34,8 +35,17 @@ class VLoss(keras.losses.Loss):
         :param batch_axis: axis used for mini-batch (default is 0) and excluded from loss aggregation.
         """
         super().__init__(**kwargs)
+        assert (loss_type == MeanSquaredError) or (loss_type == Huber), "Only expecting L2Loss or HuberLoss."
         self.loss_type = loss_type
         self.loss_fn = keras.losses.mean_squared_error#keras.losses.get(loss_type)
+
+    @property
+    def input_schema(self) -> LossInputSchema:
+        return LossInputSchema(
+            head_outputs=['pred'],
+            agent_inputs=[],
+            targets=['target']
+        )
 
     def call(self, prediction, target):
         """
