+++ rl_coach/coach.py
@@ -799,13 +799,14 @@ if __name__ == "__main__":
 
 
     sys.argv.append('-p')
-    # sys.argv.append('CartPole_DQN')
+    sys.argv.append('CartPole_DQN')
 
     # sys.argv.append('Atari_DQN')
     # sys.argv.extend(['-lvl', 'breakout'])
 
-    sys.argv.append('Mujoco_ClippedPPO')
-    sys.argv.extend(['-lvl', 'inverted_pendulum'])
+    # sys.argv.append('Mujoco_ClippedPPO')
+    # sys.argv.extend(['-lvl', 'inverted_pendulum'])
+
